# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.ExpenseCreate:
    properties:
      amount:
        type: number
      details:
        type: string
      movement_type_id:
        type: string
      purchase_order_id:
        type: string
    required:
    - amount
    - details
    - movement_type_id
    type: object
  schemas.ExpenseDTO:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      movement_type:
        $ref: '#/definitions/schemas.MovementTypeDTO'
    type: object
  schemas.ExpenseResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      details:
        type: string
      id:
        type: string
      movement_type:
        $ref: '#/definitions/schemas.MovementTypeDTO'
    type: object
  schemas.ExpenseUpdate:
    properties:
      amount:
        type: number
      details:
        type: string
      id:
        type: string
      movement_type_id:
        type: string
      purchase_order_id:
        type: string
    required:
    - amount
    - details
    - movement_type_id
    type: object
...
  schemas.MovementTypeDTO:
    properties:
      id:
        type: string
      is_income:
        type: boolean
      name:
        type: string
    type: object
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is a api to app noa gestion
  termsOfService: http://swagger.io/terms/
  title: APP NOA Gestion
  version: "1.0"
paths:
...
  /api/v1/expense/{id}:
    get:
      consumes:
      - application/json
      description: Get Expense By ID
      parameters:
      - description: ID of Expense
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expense obtained successfully
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ExpenseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Expense not found
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get Expense By ID
      tags:
      - Expense
  /api/v1/expense/create:
    post:
      consumes:
      - application/json
      description: Parses the request body to create a new expense entry.
      parameters:
      - description: Expense information
        in: body
        name: expenseCreate
        required: true
        schema:
          $ref: '#/definitions/schemas.ExpenseCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Expense created successfully
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Model Invalid
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Create Expense
      tags:
      - Expense
  /api/v1/expense/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an expense based on the provided ID and workplace context.
      parameters:
      - description: ID of the expense
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expense deleted successfully
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Delete Expense
      tags:
      - Expense
  /api/v1/expense/get_all:
    get:
      consumes:
      - application/json
      description: Fetches all expenses from the specified tenant.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of expenses
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.ExpenseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get all expenses
      tags:
      - Expense
  /api/v1/expense/get_today:
    get:
      consumes:
      - application/json
      description: Fetches all expenses from the specified tenant, on the current
        day.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of expenses
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.ExpenseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get expense today
      tags:
      - Expense
  /api/v1/expense/update:
    put:
      consumes:
      - application/json
      description: Updates the details of an expense based on the provided data.
      parameters:
      - description: Expense data to update
        in: body
        name: expenseUpdate
        required: true
        schema:
          $ref: '#/definitions/schemas.ExpenseUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Expense updated successfully
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Model Invalid
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Update Expense
      tags:
      - Expense
...
securityDefinitions:
  CookieAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"